refactoring for separate controllers
